backend/data_embedding_handler/pyproject.toml
[tool.poetry]
name = "data-embedding-handler"
version = "0.1.0"
description = "Demo handler for embedding data from source to different destinations"
authors = ["alexau <alexuau922@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
onnxruntime = "^1.16.1"
opensearch-py = "^2.3.2"
boto3-stubs = { extras = [
    "sagemaker-runtime",
    "secretsmanager",
    "sqs",
], version = "^1.28.73" }
boto3 = "^1.28.68"
transformers = { extras = ["onnx"], version = "^4.34.1" }
aws-lambda-powertools = "^2.26.0"
typing-extensions = "^4.8.0"
psycopg-binary = "^3.1.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
